Move trait to akka streams only area
